Use duplicate Sync classes in Protocol1 and Protocol3.
