+++ lib/timex_datalink_client/protocol_1/sync.rb
@@ -1,10 +1,39 @@
 # frozen_string_literal: true
 
-require "timex_datalink_client/sync"
-
 class TimexDatalinkClient
   class Protocol1
-    class Sync < Sync
+    class Sync
+      SYNC_1_BYTE = [0x55]
+      SYNC_2_BYTE = [0xaa]
+
+      SYNC_2_LENGTH = 40
+
+      attr_accessor :length
+
+      # Create a Sync instance.
+      #
+      # @param length [Integer] Number of 0x55 sync bytes to use.
+      # @return [Sync] Sync instance.
+      def initialize(length: 300)
+        @length = length
+      end
+
+      # Compile packets for syncronization data.
+      #
+      # @return [Array<Array<Integer>>] Two-dimensional array of integers that represent bytes.
+      def packets
+        [render_sync_1 + render_sync_2]
+      end
+
+      private
+
+      def render_sync_1
+        SYNC_1_BYTE * length
+      end
+
+      def render_sync_2
+        SYNC_2_BYTE * SYNC_2_LENGTH
+      end
     end
   end
 end
